清理项目：删除临时文件和文档，修复未使用变量警告
- 清理target目录（4.3GB编译中间文件）
- 删除临时报告文档（17个）
- 删除临时脚本和测试文件（11个）
- 修复未使用变量警告（queue_id, shm_id, addr_len, buf, flags等）
- 优化代码质量，减少警告数量

+++ kernel/src/ipc/signal.rs
@@ -372,7 +372,7 @@ impl SignalState {
 
         // Check if signal should be delivered to signalfd first
         let pid = crate::process::myproc().unwrap_or(0);
-        if crate::syscalls::glib::deliver_signal_to_signalfd(pid, sig, SigInfo {
+        if crate::syscalls::glib::deliver_signal_to_signalfd(pid as usize, sig, SigInfo {
             signo: sig as i32,
             code: si_code::SI_KERNEL,
             ..Default::default()
@@ -396,7 +396,7 @@ impl SignalState {
 
         // Check if signal should be delivered to signalfd first
         let pid = crate::process::myproc().unwrap_or(0);
-        if crate::syscalls::glib::deliver_signal_to_signalfd(pid, sig, info) {
+        if crate::syscalls::glib::deliver_signal_to_signalfd(pid as usize, sig, info) {
             // Signal was delivered to signalfd, don't set pending bit
             return Ok(());
         }
