+++ kernel/src/sync/futex_tests.rs
@@ -5,9 +5,6 @@
 //! and performance benchmarks.
 
 extern crate alloc;
-
-use alloc::vec::Vec;
-use alloc::collections::BTreeMap;
 use core::sync::atomic::{AtomicI32, AtomicUsize, Ordering};
 use crate::sync::Mutex;
 use crate::syscalls::thread::{
@@ -335,8 +332,8 @@ impl FutexTestSuite {
         let avg_latency = total_time / operations;
         
         // Update stats
-        self.stats.total_operations += operations;
-        self.stats.successful_operations += operations;
+        self.stats.total_operations += operations as usize;
+        self.stats.successful_operations += operations as usize;
         self.stats.avg_latency_ns = avg_latency;
         
         crate::println!("[futex_test] Performance test passed, avg latency: {}ns", avg_latency);
